// src/app/model/risorse.ts
export interface Risorse {
  id: number;
  nomeCognome: string;
  email: string;
  recapito: string;
  profiloLinkedin: string;
  citta: string;
  competenzaPrincipale: string;
  dataColloquio: Date;
  annoColloquio: number;
  fonteReperimento: string;
  costoGiornaliero: number;
  possibilitaLavorativa: string;
  skillCampoLibero: string;
  competenzeTotali: string;
  certificazioni: string;
}