// src/app/model/dettagli_risorse.ts
export interface DettagliRisorse {
  id: number;
  idRisorsa: number;
  fileBase64: string;
  dataInserimento: Date;
  idEsitoColloquio: number;
  idProfilo: number;
  idSkill1: number;
  idSkill2: number;
  idSkill3: number;
  idSkill4: number;
  idSkill5: number;
  idLingua1: number;
  idLingua2: number;
  idLingua3: number;
  idSeniority: number;
}